Added test workflow against a mocked server.
Refs #17
