+++ tests/META-INF/MANIFEST.MF
@@ -6,10 +6,13 @@ Bundle-Version: 1.0.13
 Bundle-RequiredExecutionEnvironment: JavaSE-1.7
 Require-Bundle: org.junit;bundle-version="4.12.0",
  nl.esciencecenter.e3dchem.gpcrdb,
+ nl.esciencecenter.e3dchem.knime.testing.plugin;bundle-version="[1.0.0,2.0.0)",
  org.knime.core;bundle-version="[3.1.0,4.0.0)",
  org.knime.json;bundle-version="[3.1.0,4.0.0)",
+ org.knime.testing;bundle-version="[3.1.0,4.0.0)",
  org.knime.base;bundle-version="[3.1.0,4.0.0)"
 Bundle-Vendor: Netherlands eScience Center
 Bundle-ClassPath: lib/mockito-all-1.10.19.jar,
  lib/jackson-core-2.4.2.jar,
+ lib/wiremock-standalone-2.1.12.jar,
  .
