+++ tests/build.properties
@@ -1,6 +1,7 @@
-source.. = src/
+source.. = src/java
 output.. = target/classes/
 bin.includes = META-INF/,\
                .,\
                lib/mockito-all-1.10.19.jar,\
-               lib/jackson-core-2.4.2.jar
+               lib/jackson-core-2.4.2.jar,\
+               lib/wiremock-standalone-2.1.12.jar
