+++ tests/src/java/nl/esciencecenter/e3dchem/gpcrdb/NodePluginTest.java
@@ -1,0 +1,13 @@
+package nl.esciencecenter.e3dchem.gpcrdb;
+
+import static org.junit.Assert.assertEquals;
+
+import org.junit.Test;
+
+public class NodePluginTest {
+
+	@Test
+	public void testIt() {
+		assertEquals(1, 1);
+	}
+}
