+++ tests/src/knime/protein-families-test/.savedWithData
@@ -1,0 +1,4 @@
+Do not delete this file!
+This file serves to indicate that the workflow was written as part of the usual save routine (not exported).
+
+Workflow was last saved by user vagrant on Wed Sep 28 14:55:53 UTC 2016
