+++ tests/src/knime/workflowset.meta
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="UTF-8"?><KNIMEMetaInfo nrOfElements="2">
+<element form="text" name="Author"/>
+<element form="date" name="Creation Date">28/8/2016</element>
+<element form="multiline" name="Comments"/>
+</KNIMEMetaInfo>
