+++ tests/src/test/resources/mappings/mapping-services-proteinfamily-iO9Fk.json
@@ -1,0 +1,22 @@
+{
+  "uuid" : "fb672721-3409-3984-89e2-8bd6fcefc871",
+  "request" : {
+    "url" : "/services/proteinfamily/",
+    "method" : "GET"
+  },
+  "response" : {
+    "status" : 200,
+    "bodyFileName" : "body-services-proteinfamily-iO9Fk.json",
+    "headers" : {
+      "Date" : "Wed, 28 Sep 2016 14:46:39 GMT",
+      "Server" : "Apache/2.4.7 (Ubuntu)",
+      "Allow" : "GET, HEAD, OPTIONS",
+      "X-Frame-Options" : "SAMEORIGIN",
+      "Vary" : "Accept,Cookie",
+      "Keep-Alive" : "timeout=5, max=100",
+      "Connection" : "Keep-Alive",
+      "Transfer-Encoding" : "chunked",
+      "Content-Type" : "application/json"
+    }
+  }
+}
